Demo test for contract throwing
When a contract throws, you get an invalid jump error. This shows
how such an error can be looked for in truffle.

Note that these are bad tests, as they are not independent (i.e. running
some of them with it.only fails)! This should be cleared up at
some point.