Use geth for testing on Circle
Our contracts behaving correctly on `geth` is what we actually care
about, so let's test using it on Circle. We cache making the
DAG so that we don't have to do that every time.

A `genesis.json` file has been added to ./truffle/ which is used
by `geth` to start the blockchain. The strings `xxxxx` and
`yyyyy` are replaced during the test setup on Circle with the
first two addresses generated, to ensure that we have ether to
run the first tests. The coinbase is set to the third address
so that the ether balances aren't messed up as blocks are
generated.

Also a slight refactoring of some tests for the benefit of
bithound; it does also make the code read more nicely.