Add colony switching to POC
Also includes - by necessity - colony creation. Note that for
any colony specific stuff to work now, Web3Manager.changeColony
has to be called to set colonyDeployed to be using the right
colony. Only the RootColony is set up by default now, which
allow creation and browsing a list of colonies.

Is this drifting away from how we should be doing it in React?
Should I set the address of the components, and then the components
pass the address as an argument to the constructor of colonyJS
(replacing the Web3Manager.deployed() argument)
that then sets everything up correctly? Basically, I think I've
drunk the React Kool Aid and I'm now paranoid about state.