Fix waiting for blockchain in test setups
In these cases we weren't returning the promises that pudding
generated for us, so the rest of the promise chain wasn't able
to wait for the transactions to be mined. This made the tests
intermittent, but mostly only on circle which uses Geth for these
tests with longer mining times.

We add an extra function wrapping the done() call so we don't get
Error: done() invoked with non-Error by calling done() with
the transaction ID after the fact.