Fix issue with mining long logs
Previously, we called delete on an array in ColonyNetwork, but I was
mistaken in thinking that that took constant gas - it actually assigns
to every element in the whole array, which obviously runs in to issues
with gas. The only way to delete data with constant gas costs is
`selfdestruct`. Fortunately, we already had a related contract
(ReputationMiningCycle) that we `selfdestruct`. This commit moves
the storage of the reputation log to the ReputationMiningCycle
contract, allowing it to be cleaned up nicely.

I think there is a gas penalty for this implementation, but it's better
than having reputation mining lock up once more than 20 reputation updates
happen in an hour.