Remove requires for MAX_INT values
Any value that is used is proved to be in the reputation tree; the
correct state cannot have such a value, and so when a submission
comes up against the correct reputation, they won't be able to prove
such values are in the agree state. They will be able to prove the new
value is in the disagree state, if they so choose, but then the
calculation will fail.