Use EthGasStation API for gas prices
The API is deprecated though, so we should probably keep
an eye on it....

For transactions where there isn't a rush, 'safeLow' is used
and for transactions where there is (i.e. disputes), then 'fast' is
used. In the future, the original submission probably wants to be 'fast'
as well. We may even provide users an option to use 'fastest' if they
want...