Never restart if miner out of gas
If we restart when the miner is out of gas, we end up in a restart
loop, because it's almost certainly still out of gas a few seconds
later. This PR means that it just sits patiently until it is
manually restarted (presumably, after it has been sent some ether).